fixes #5
