+++ README.md
@@ -7,7 +7,7 @@ Install using pip (>= 19.0).
 
 ```bash
 pip install --upgrade pip
-pip install git+https://github.com/cms-l1-globaltrigger/tm-diff.git@0.7.2
+pip install git+https://github.com/cms-l1-globaltrigger/tm-diff.git@0.7.3
 ```
 
 ## Basic usage
