+++ requirements.txt
@@ -1,1 +1,1 @@
-tm-python @ git+https://github.com/cms-l1-globaltrigger/tm-python@0.9.1
+tm-python @ git+https://github.com/cms-l1-globaltrigger/tm-python@0.10.0
