+++ tmDiff/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '0.7.2'
+__version__ = '0.7.3'
